requirements.txt
altair==4.2.2
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
decorator==5.1.1
entrypoints==0.4
frozendict==2.3.7
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-pubsub==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
html5lib==1.1
idna==3.4
importlib-metadata==6.5.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multitasking==0.0.11
numpy==1.24.2
packaging==23.1
pandas==1.5.3
patsy==0.5.3
Pillow==9.5.0
plotly==5.14.1
plotly-express==0.4.1
polygon-api-client==1.8.5
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.4
rsa==4.9
scipy==1.10.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4
statsmodels==0.13.5
streamlit==1.21.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
websockets==10.4
yfinance==0.2.18
zipp==3.15.0
